Update ECalBarrelInclined_geo.cpp
A patch has been added to the ```ECalBarrelInclined_geo.cpp``` file so that it can now access the information like Radiation length and Interaction length using the ```MaterialManager```. The values for ```sensitive_thickness, inner_thickness, outer_thickness and absorberThickness``` are temporarily dummy. They will be changed soon and proper values will be added.